+++ models/src/main/java/io/sphere/sdk/products/commands/updateactions/RemovePrice.java
@@ -8,10 +8,10 @@ import io.sphere.sdk.products.ProductUpdateScope;
  * {@include.example io.sphere.sdk.products.ProductCrudIntegrationTest#removePriceUpdateAction()}
  */
 public class RemovePrice extends StageableProductUpdateAction {
-    private final long variantId;
+    private final int variantId;
     private final Price price;
 
-    private RemovePrice(final long variantId, final Price price, final ProductUpdateScope productUpdateScope) {
+    private RemovePrice(final int variantId, final Price price, final ProductUpdateScope productUpdateScope) {
         super("removePrice", productUpdateScope);
         this.variantId = variantId;
         this.price = price;
@@ -25,7 +25,7 @@ public class RemovePrice extends StageableProductUpdateAction {
         return price;
     }
 
-    public static RemovePrice of(final long variantId, final Price price, final ProductUpdateScope productUpdateScope) {
+    public static RemovePrice of(final int variantId, final Price price, final ProductUpdateScope productUpdateScope) {
         return new RemovePrice(variantId, price, productUpdateScope);
     }
 }
